refactor: 프로젝트 폴더 구조 정리 및 역할별 파일 분리
- components/: 재사용 가능한 UI 컴포넌트를 관리
- constants/: 상수 값을 한 곳에서 관리
- data/: 레스토랑 데이터를 관리
- handlers/: 이벤트 핸들러를 모아서 관리
- store/: 상태 관리를 위한 store 파일을 별도로 관리
- styles/: 스타일 파일을 분리
